Object now show on screen
It does not appear when the window is loaded (only when mouse is clicked and moved). Object is all black. The World and Object Bound are different (even though the ObjectToWorld transformation should be identity, and the World bound is just the Object bound transformed by ObjectToWorld). Object does not appear in the center of the world. Camera still iffy.