Teapot now renders without hole.
The (inelegant) patchwork solution is to keep a separate quads list, in addition to the triangle list. However, the only other alternative at the moment is to manually convert quads to triangle, which for some reason does not work (if we change GL_QUADS to GL_TRIANGLE_STRIP it also does not work), and bloat up the size anyway.